reworking the motion profiler
