+++ include/Constants/ConstantsReal.h
@@ -4,7 +4,6 @@
 
 const float azimuthMaxVelocity      = 10.0; // degrees per second
 const float azimuthMaxAcceleration  = 10.0; // degrees per second^2
-const float azimuthMaxJerk          = 10.0; // degrees per second^3
 
 const float azimuthAcceptableError = 0.0; // degrees
 const float azimuthAcceptableVelocityError = 0.0; // degrees per second
@@ -16,7 +15,6 @@ const float azimuthkD   = 0.0; // unitless
 
 const float elevationMaxVelocity      = 10.0; // degrees per second
 const float elevationMaxAcceleration  = 10.0; // degrees per second^2
-const float elevationMaxJerk          = 10.0; // degrees per second^3
 
 const float elevationAcceptableError = 0.0; // degrees
 const float elevationAcceptableVelocityError = 0.0; // degrees per second
